Struct jconfig::config::Config

pub struct Config {
    // some fields omitted
}

A map of the configuration entries

Methods

impl Config

fn new() -> Self

Creates a new, empty configuration

fn from_str<'a>(s: &'a str) -> Result<Self>

Load a Config from a JSON string

fn load<P: AsRef<Path>>(path: P) -> Result<Self>

Load a Config from a file

fn set<V: Field>(&mut self, value: V)

Set a config field to the corresponding value.

The field is determined by the type of the value being set.

fn get_raw(&self, name: &str) -> Option<RawValue>

Access the raw value of a field.

Prefer to use the typed getters instead.

fn set_raw<K: Into<Cow<'static, str>>>(&mut self, name: K, value: RawValue)

Set the raw value of a field, bypassing any typed fields.

fn remove_raw(&mut self, name: &str)

Remove a field by name

fn get<V: Field>(&self) -> Option<&V>

Get a reference to a field's value, if it exists.

fn get_mut<V: Field>(&mut self) -> Option<&mut V>

Get a mutable reference to a field's value, if it exists.

fn has<V: Field>(&self) -> bool

Returns a boolean of whether a certain field is in the config.

fn remove<V: Field>(&mut self) -> bool

Removes a field from the config, if one existed. Returns true if field has been removed.

fn iter(&self) -> ConfigItems

Returns an iterator over the fields.

fn len(&self) -> usize

Returns the number of fields in the config.

fn clear(&mut self)

Remove all fields from the config.