Trait jconfig::field::FieldRef

pub trait FieldRef {
    fn get<V>(&self) -> Option<&V> where V: Field;
    fn get_raw(&self, name: &str) -> Option<RawValue>;
}

Trait for viewing the contents of a Field structure.

Required Methods

fn get<V>(&self) -> Option<&V> where V: Field

View a reference to a field if it exists.

fn get_raw(&self, name: &str) -> Option<RawValue>

View a reference to the RawValue of the field if it exists.